vastly improved PDF system
This probably won't be the final form of it. I may rethink the core 
rendering munger just a bit, and see if it can become more extensible. 
That way any verb could output content for mpdf and mark it to be 
rendered as a PDF rather than an HTML page.